Extracted ManageReposModal from PluginBrowse
